wip: ChatVertexAI.with_structured_output
