+++ libs/langchain/langchain/chains/openai_functions/extraction.py
@@ -1,6 +1,10 @@
 from typing import Any, List, Optional
 
 from langchain_core.language_models import BaseLanguageModel
+from langchain_core.output_parsers.openai_functions import (
+    JsonKeyOutputFunctionsParser,
+    PydanticAttrOutputFunctionsParser,
+)
 from langchain_core.prompts import BasePromptTemplate, ChatPromptTemplate
 from langchain_core.pydantic_v1 import BaseModel
 
@@ -10,10 +14,6 @@ from langchain.chains.openai_functions.utils import (
     _convert_schema,
     _resolve_schema_references,
     get_llm_kwargs,
-)
-from langchain.output_parsers.openai_functions import (
-    JsonKeyOutputFunctionsParser,
-    PydanticAttrOutputFunctionsParser,
 )
