+++ libs/langchain/langchain/chains/openai_functions/qa_with_structure.py
@@ -3,16 +3,16 @@ from typing import Any, List, Optional, Type, Union
 from langchain_core.language_models import BaseLanguageModel
 from langchain_core.messages import HumanMessage, SystemMessage
 from langchain_core.output_parsers import BaseLLMOutputParser
+from langchain_core.output_parsers.openai_functions import (
+    OutputFunctionsParser,
+    PydanticOutputFunctionsParser,
+)
 from langchain_core.prompts import PromptTemplate
 from langchain_core.prompts.chat import ChatPromptTemplate, HumanMessagePromptTemplate
 from langchain_core.pydantic_v1 import BaseModel, Field
 
 from langchain.chains.llm import LLMChain
 from langchain.chains.openai_functions.utils import get_llm_kwargs
-from langchain.output_parsers.openai_functions import (
-    OutputFunctionsParser,
-    PydanticOutputFunctionsParser,
-)
 
 
 class AnswerWithSources(BaseModel):
