+++ libs/langchain/langchain/chains/openai_functions/tagging.py
@@ -1,15 +1,15 @@
 from typing import Any, Optional
 
 from langchain_core.language_models import BaseLanguageModel
+from langchain_core.output_parsers.openai_functions import (
+    JsonOutputFunctionsParser,
+    PydanticOutputFunctionsParser,
+)
 from langchain_core.prompts import ChatPromptTemplate
 
 from langchain.chains.base import Chain
 from langchain.chains.llm import LLMChain
 from langchain.chains.openai_functions.utils import _convert_schema, get_llm_kwargs
-from langchain.output_parsers.openai_functions import (
-    JsonOutputFunctionsParser,
-    PydanticOutputFunctionsParser,
-)
 
 
 def _get_tagging_function(schema: dict) -> dict:
