+++ libs/langchain/langchain/runnables/openai_functions.py
@@ -2,11 +2,10 @@ from operator import itemgetter
 from typing import Any, Callable, List, Mapping, Optional, Union
 
 from langchain_core.messages import BaseMessage
+from langchain_core.output_parsers.openai_functions import JsonOutputFunctionsParser
 from langchain_core.runnables import RouterRunnable, Runnable
 from langchain_core.runnables.base import RunnableBindingBase
 from typing_extensions import TypedDict
-
-from langchain.output_parsers.openai_functions import JsonOutputFunctionsParser
 
 
 class OpenAIFunction(TypedDict):
