+++ libs/langchain/tests/unit_tests/output_parsers/test_openai_functions.py
@@ -4,12 +4,12 @@ from typing import Any, Dict
 import pytest
 from langchain_core.exceptions import OutputParserException
 from langchain_core.messages import AIMessage, BaseMessage, HumanMessage
-from langchain_core.outputs import ChatGeneration
-
-from langchain.output_parsers.openai_functions import (
+from langchain_core.output_parsers.openai_functions import (
     JsonOutputFunctionsParser,
     PydanticOutputFunctionsParser,
 )
+from langchain_core.outputs import ChatGeneration
+
 from langchain.pydantic_v1 import BaseModel
