+++ templates/chat-bot-feedback/chat_bot_feedback/chain.py
@@ -5,7 +5,6 @@ from typing import List, Optional
 from langchain import hub
 from langchain.callbacks.tracers.evaluation import EvaluatorCallbackHandler
 from langchain.callbacks.tracers.schemas import Run
-from langchain.output_parsers.openai_functions import JsonOutputFunctionsParser
 from langchain.schema import (
     AIMessage,
     BaseMessage,
@@ -14,6 +13,7 @@ from langchain.schema import (
     get_buffer_string,
 )
 from langchain_community.chat_models import ChatOpenAI
+from langchain_core.output_parsers.openai_functions import JsonOutputFunctionsParser
 from langchain_core.prompts import ChatPromptTemplate, MessagesPlaceholder
 from langchain_core.pydantic_v1 import BaseModel, Field
 from langchain_core.runnables import Runnable
