+++ templates/rag-multi-index-router/rag_multi_index_router/chain.py
@@ -1,7 +1,6 @@
 from operator import itemgetter
 from typing import Literal
 
-from langchain.output_parsers.openai_functions import PydanticAttrOutputFunctionsParser
 from langchain.retrievers import (
     ArxivRetriever,
     KayAiRetriever,
@@ -11,6 +10,9 @@ from langchain.retrievers import (
 from langchain.utils.openai_functions import convert_pydantic_to_openai_function
 from langchain_community.chat_models import ChatOpenAI
 from langchain_core.output_parsers import StrOutputParser
+from langchain_core.output_parsers.openai_functions import (
+    PydanticAttrOutputFunctionsParser,
+)
 from langchain_core.prompts import ChatPromptTemplate
 from langchain_core.pydantic_v1 import BaseModel, Field
 from langchain_core.runnables import (
